move common command behaviour to own package
