+++ pkg/exec/doc.go
@@ -1,0 +1,2 @@
+// Package exec contains helper abstractions for working with external commands.
+package exec
